Fix of ft_num

+++ ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: biphuyal <biphuyal@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/23 17:24:36 by biphuyal          #+#    #+#             */
-/*   Updated: 2025/06/23 20:23:17 by biphuyal         ###   ########.fr       */
+/*   Updated: 2025/07/01 15:01:48 by biphuyal         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ int ft_formater(va_list args, const char format)
 	else if (format == 'p')
 		length += ft_hexa_pointer(va_arg(args, unsigned long long));
 	else if (format == 'd' || format == 'i')
-		length += ft_num(va_arg(args, int));
+		length += ft_num(va_arg(args, long long));
 	else if (format == '%')
 		length += ft_char('%');
 	else if (format == 'u')
