+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: biphuyal <biphuyal@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/23 17:24:41 by biphuyal          #+#    #+#             */
-/*   Updated: 2025/06/23 19:46:36 by biphuyal         ###   ########.fr       */
+/*   Updated: 2025/07/01 15:02:23 by biphuyal         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
 int ft_strs(char *str);
 int ft_char(int c);
 int ft_unsigned(unsigned int n);
-int ft_num(int n);
+int ft_num(long long n);
 int ft_hexa_lower(unsigned int n);
 int ft_hexa_upper(unsigned int n);
 int ft_hexa_pointer(unsigned long long n);
